Fix build of grunt_site

+++ flipside_platform/build/grunt_build.py
@@ -11,5 +11,6 @@ from .. import config
 def build(dist_dir, environment, **kwargs):
     subprocess.check_call(['grunt', 'build:{}'.format(environment)])
     frontend_dir = os.path.join(dist_dir, 'frontend')
-    shutil.rmtree(frontend_dir)
+    if os.path.exists(frontend_dir):
+        shutil.rmtree(frontend_dir)
     shutil.copytree('dist', frontend_dir)
